+++ lib/feature/login/ui/chickBox.dart
@@ -1,0 +1,44 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+class ChickBox extends StatefulWidget {
+  const ChickBox({super.key});
+
+  @override
+  State<ChickBox> createState() => _ChickBoxState();
+}
+
+class _ChickBoxState extends State<ChickBox> {
+  bool isChecked = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Row(
+        children: [
+          Checkbox(
+            value: isChecked,
+            onChanged: (bool? newValue) {
+              setState(() {
+                isChecked = newValue ?? false;
+              });
+            },
+          ),
+          Text(
+            "Remember password",
+            style: TextStyle(fontSize: 13.sp, fontWeight: FontWeight.w400),
+          ),
+          Spacer(),
+          TextButton(
+            onPressed: () {},
+            child: Text(
+              "Forgot Password",
+              style: TextStyle(fontSize: 12.sp, fontWeight: FontWeight.w400),
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
